Add mappings file
Mappings file generated using Dumper-7. Useful for example in FModel. 

If you want to generate this file for yourself, you'll need to compile the Dumper-7 DLL yourself and then use System Informer (previously known as Process Hacker) to load it. 
You do that by going into the properties of the Half Sword "shipping" EXE, going to the Modules tab and loading the DLL as a module there. It will create a folder on your C:\ drive. The mappings file will be in there.
If you want this folder to be placed elsewhere, you'll need to recompile the DLL with your desired path set. Dumper-7's GitHub repo has better instructions for this process.